requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
typing-extensions>=4.5.0
anyio>=4.0.0
email-validator>=2.0.0
httpx>=0.27.0
supabase>=2.4.0